move board rule into board module
