+++ apps/server/src/modules/authorization/domain/rules/index.ts
@@ -2,7 +2,6 @@
  * Rules are currently placed in authorization module to avoid dependency cycles.
  * In future they must be moved to the feature modules and register it in registration service.
  */
-export * from './board-node.rule';
 export * from './context-external-tool.rule';
 export * from './course-group.rule';
 export * from './course.rule';
